add security layer with jwt

+++ Task9_SecurityExpand/src/main/java/com/example/task9_securityexpand/model/Product.java
@@ -1,16 +1,14 @@
 package com.example.task9_securityexpand.model;
 
 import com.example.task9_securityexpand.currency.Currency;
-import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import javax.persistence.*;
 import java.io.Serializable;
 import java.math.BigDecimal;
-import java.time.LocalDate;
-
 import java.util.List;
 
 @Data
@@ -22,7 +20,7 @@ import java.util.List;
 public class Product implements Serializable{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     private String name;
     private String description;
     @Enumerated(EnumType.STRING)
